Turn `start_daemon` into more reusable code
